Fast average spectrum evaluation
