+++ src/metrics.py
@@ -1,42 +1,122 @@
-import numpy as np
 import sys
+from typing import Iterable
+
+import numba as nb
+import numpy as np
 import spectrum_utils.spectrum as sus
-from pyteomics import mgf, parser
-from scipy.stats import binned_statistic
+from pyteomics import parser
+
 
 mz_unit = 1.000508  # Space in Th between fragments
 mz_space = mz_unit * .005  # Resolution approximately 0.005 Da
 
 
-def bin_proc(spectrum, mz_space, max_mz):
-    bins = np.arange(-mz_space / 2., max_mz, mz_space)
-    # bins = np.linspace(-mz_space/2., max_mz, num=xxx)
-    dig_spec, _, __ = binned_statistic(spectrum.mz, spectrum.intensity,
-                                       statistic='sum', bins=bins)
-    return dig_spec
+def dot(spectrum1: sus.MsmsSpectrum, spectrum2: sus.MsmsSpectrum,
+        fragment_mz_tolerance: float) -> float:
+    """
+    Compute the dot product between the given spectra.
+
+    Parameters
+    ----------
+    spectrum1 : sus.MsmsSpectrum
+        The first spectrum.
+    spectrum2 : sus.MsmsSpectrum
+        The second spectrum.
+    fragment_mz_tolerance : float
+        The fragment m/z tolerance used to match peaks.
+
+    Returns
+    -------
+    float
+        The dot product similarity between the given spectra.
+    """
+    return _dot(spectrum1.mz, _norm_intensity(np.copy(spectrum1.intensity)),
+                spectrum2.mz, _norm_intensity(np.copy(spectrum2.intensity)),
+                fragment_mz_tolerance)
 
 
-def cos_dist(representative_spectrum, cluster_member):
-    max_mz = max(representative_spectrum.mz[-1], cluster_member.mz[-1])
-    discrete_a = bin_proc(representative_spectrum, mz_space, max_mz)
-    discrete_b = bin_proc(cluster_member, mz_space, max_mz)
-    a = np.dot(discrete_a, discrete_a)
-    b = np.dot(discrete_b, discrete_b)
-    ab = np.dot(discrete_a, discrete_b)
-    if a == 0. or b == 0.:
-        return 0.
-    else:
-        return ab / np.sqrt(a * b)
+@nb.njit
+def _norm_intensity(spectrum_intensity: np.ndarray) -> np.ndarray:
+    """
+    Normalize spectrum peak intensities.
+
+    Parameters
+    ----------
+    spectrum_intensity : np.ndarray
+        The spectrum peak intensities to be normalized.
+
+    Returns
+    -------
+    np.ndarray
+        The normalized peak intensities.
+    """
+    return spectrum_intensity / np.linalg.norm(spectrum_intensity)
 
 
-def average_cos_dist(representative_spectrum, cluster_members):
-    sum_dist = 0.0
-    for member in cluster_members:
-        sum_dist += cos_dist(representative_spectrum, member)
-    if len(cluster_members) > 0:
-        return sum_dist / float(len(cluster_members))
-    else:
-        return 0.0
+@nb.njit
+def _dot(mz: np.ndarray, intensity: np.ndarray, mz_other: np.ndarray,
+         intensity_other: np.ndarray, fragment_mz_tol: float) -> float:
+    """
+    Compute the dot product between the given spectra.
+
+    Note: Spectrum intensities should be normalized prior to computing the dot
+    product.
+
+    Parameters
+    ----------
+    mz : np.ndarray
+        The first spectrum's m/z values.
+    intensity : np.ndarray
+        The first spectrum's intensity values.
+    mz_other : np.ndarray
+        The second spectrum's m/z values.
+    intensity_other : np.ndarray
+        The second spectrum's intensity values.
+    fragment_mz_tol : float
+        The fragment m/z tolerance used to match peaks in both spectra with
+        each other.
+
+    Returns
+    -------
+    float
+        The dot product between both spectra.
+    """
+    fragment_i, fragment_other_i, score = 0, 0, 0.
+    for fragment_i in range(len(mz)):
+        while (fragment_other_i < len(mz_other) - 1 and
+               mz_other[fragment_other_i] < mz[fragment_i] - fragment_mz_tol):
+            fragment_other_i += 1
+        if (abs(mz[fragment_i] - mz_other[fragment_other_i]) <= fragment_mz_tol
+                and fragment_other_i < len(mz_other)):
+            score += intensity[fragment_i] * intensity_other[fragment_other_i]
+            fragment_other_i += 1
+    return score
+
+
+def avg_dot(representative: sus.MsmsSpectrum,
+            cluster_spectra: Iterable[sus.MsmsSpectrum],
+            fragment_mz_tolerance: float) -> float:
+    """
+    Compute the average dot product between the cluster representative and all
+    cluster members.
+
+    Parameters
+    ----------
+    representative : sus.MsmsSpectrum
+        The cluster representative spectrum.
+    cluster_spectra : Iterable[sus.MsmsSpectrum]
+        The cluster member spectra.
+    fragment_mz_tolerance : float
+        Fragment m/z tolerance used during spectrum comparison.
+
+    Returns
+    -------
+    float
+        The average dot product between the cluster representative and all
+        cluster members.
+    """
+    return np.mean([dot(representative, spectrum, fragment_mz_tolerance)
+                    for spectrum in cluster_spectra])
 
 
 def fraction_of_by(representative_spectrum, cluster_members=[]):
@@ -66,25 +146,3 @@ def fraction_of_by_seq(peptide_seq, precursor_mz, precursor_charge, mz, intensit
         peptide_seq, precursor_mz=precursor_mz, precursor_charge=precursor_charge, mz=mz, intensity=intensity,
         peptide=peptide_seq)
     return fraction_of_by(spec)
-
-
-if __name__ == "__main__":
-    # If the library is called as main, run some rudementary tests of the functions
-
-    for spectrum_dict in mgf.read("datasets/clusters_maracluster.mgf"):
-        #    for spectrum_dict in mgf.read("datasets/.mgf"):
-        peptide_seq = spectrum_dict['params']['title'].split(':')[-1][:-2]
-        precursor_mz = spectrum_dict['params']['pepmass'][0]
-        precursor_charge = spectrum_dict['params']['charge'][0]
-        mz = spectrum_dict['m/z array']
-        intensity = spectrum_dict['intensity array']
-        break
-
-    print(fraction_of_by_seq(peptide_seq, precursor_mz, precursor_charge, mz, intensity))
-
-    spec = sus.MsmsSpectrum(
-        peptide_seq, precursor_mz=precursor_mz, precursor_charge=precursor_charge, mz=mz, intensity=intensity,
-        peptide=None)
-
-    print(average_cos_dist(spec, [spec]))
-    print(fraction_of_by(spec, []))
