+++ src/selector.py
@@ -1,13 +1,13 @@
 import abc
 import copy
+import functools
 import math
 from typing import Dict
 
-import numba as nb
 import numpy as np
-import pyopenms
 import spectrum_utils.spectrum as sus
 
+import metrics
 import ms_io
 
 
@@ -107,88 +107,6 @@ class BestSpectrumRepresentativeSelector(RepresentativeSelector):
                                    else scores.idxmin()]
 
 
-def dot(spectrum1: sus.MsmsSpectrum, spectrum2: sus.MsmsSpectrum,
-        fragment_mz_tolerance: float) -> float:
-    """
-    Compute the dot product between the given spectra.
-
-    Parameters
-    ----------
-    spectrum1 : sus.MsmsSpectrum
-        The first spectrum.
-    spectrum2 : sus.MsmsSpectrum
-        The second spectrum.
-    fragment_mz_tolerance : float
-        The fragment m/z tolerance used to match peaks.
-
-    Returns
-    -------
-    float
-        The dot product similarity between the given spectra.
-    """
-    return _dot(spectrum1.mz, _norm_intensity(np.copy(spectrum1.intensity)),
-                spectrum2.mz, _norm_intensity(np.copy(spectrum2.intensity)),
-                fragment_mz_tolerance)
-
-
-@nb.njit
-def _norm_intensity(spectrum_intensity: np.ndarray) -> np.ndarray:
-    """
-    Normalize spectrum peak intensities.
-
-    Parameters
-    ----------
-    spectrum_intensity : np.ndarray
-        The spectrum peak intensities to be normalized.
-
-    Returns
-    -------
-    np.ndarray
-        The normalized peak intensities.
-    """
-    return spectrum_intensity / np.linalg.norm(spectrum_intensity)
-
-
-@nb.njit
-def _dot(mz: np.ndarray, intensity: np.ndarray, mz_other: np.ndarray,
-         intensity_other: np.ndarray, fragment_mz_tol: float) -> float:
-    """
-    Compute the dot product between the given spectra.
-
-    Note: Spectrum intensities should be normalized prior to computing the dot
-    product.
-
-    Parameters
-    ----------
-    mz : np.ndarray
-        The first spectrum's m/z values.
-    intensity : np.ndarray
-        The first spectrum's intensity values.
-    mz_other : np.ndarray
-        The second spectrum's m/z values.
-    intensity_other : np.ndarray
-        The second spectrum's intensity values.
-    fragment_mz_tol : float
-        The fragment m/z tolerance used to match peaks in both spectra with
-        each other.
-
-    Returns
-    -------
-    float
-        The dot product between both spectra.
-    """
-    fragment_i, fragment_other_i, score = 0, 0, 0.
-    for fragment_i in range(len(mz)):
-        while (fragment_other_i < len(mz_other) - 1 and
-               mz_other[fragment_other_i] < mz[fragment_i] - fragment_mz_tol):
-            fragment_other_i += 1
-        if (abs(mz[fragment_i] - mz_other[fragment_other_i]) <= fragment_mz_tol
-                and fragment_other_i < len(mz_other)):
-            score += intensity[fragment_i] * intensity_other[fragment_other_i]
-            fragment_other_i += 1
-    return score
-
-
 class MostSimilarRepresentativeSelector(RepresentativeSelector):
     """
     Select as cluster representative the most similar spectrum that has the
@@ -212,10 +130,10 @@ class MostSimilarRepresentativeSelector(RepresentativeSelector):
         """
         self.sim = sim
         if self.sim == 'dot':
-            self.compare_spectra = dot
+            self.compare_spectra = functools.partial(
+                metrics.dot, fragment_mz_tolerance=fragment_mz_tolerance)
         else:
             raise ValueError('Unknown spectrum similarity method')
-        self.fragment_mz_tolerance = fragment_mz_tolerance
 
     def get_description(self):
         """
@@ -256,8 +174,7 @@ class MostSimilarRepresentativeSelector(RepresentativeSelector):
             for j in range(i, len(spectra_keys)):
                 sim_matrix[i, j] = sim_matrix[j, i] = self.compare_spectra(
                     cluster_spectra[spectra_keys[i]],
-                    cluster_spectra[spectra_keys[j]],
-                    self.fragment_mz_tolerance)
+                    cluster_spectra[spectra_keys[j]])
         # Find the spectrum with the maximum similarity to all other spectra.
         return cluster_spectra[spectra_keys[sim_matrix.sum(axis=0).argmax()]]
